+++ src/coverage.ts
@@ -76,12 +76,12 @@ function toTable(coverageArr: CoverageLine[], options: Options): string {
         return changedFiles?.all.some((c) => c.includes(line.file))
       })
       // Filter folders without files
-      .filter((_line, _i, arr) => {
+      .filter((line, _i, arr) => {
         if (!reportOnlyChangedFiles) {
           return true
         }
 
-        return arr.length > 1
+        return isFile(line) || arr.some(isFile)
       })
       .map((line) => toRow(line, isFile(line), options))
     rows.push(...files)