+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Changes
 
 - fix broken links to files in coverage report when running Jest with `--changedSince` or `--findRelatedTests` (#89)
+- fix empty coverage report when `report-only-changed-files` is enabled and the report contains a single changed file
 
 ## [Jest Coverage Comment 1.0.34](https://github.com/MishaKav/jest-coverage-comment/tree/v1.0.34)